pyproject.toml
[tool.poetry]
name = "pychat"
version = "0.1.0"
description = ""
authors = ["Thiago Veronezi <tveronezi@gmail.com>"]
readme = "README.md"
packages = [{include = "pychat"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
alembic = "^1.8.0"
psycopg2 = "^2.9.3"
strawberry-graphql = {extras = ["fastapi"], version = "^0.114.3"}
SQLAlchemy = "^1.4.39"
uuid = "^1.30"

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.17.6"}
black = "^22.3.0"
mypy = "^0.961"
isort = "^5.10.1"
flake8 = "^4.0.1"
alembic = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "pychat.main:start"
